Working feide integration, and connected redux
