+++ gui/app/reducers/auth/index.js
@@ -1,72 +1,21 @@
-const LOAD = 'musit/auth/LOAD';
-const LOAD_SUCCESS = 'musit/auth/LOAD_SUCCESS';
-const LOAD_FAIL = 'musit/auth/LOAD_FAIL';
-const LOGIN = 'musit/auth/LOGIN';
-const LOGIN_SUCCESS = 'musit/auth/LOGIN_SUCCESS';
-const LOGIN_FAIL = 'musit/auth/LOGIN_FAIL';
-const LOGOUT = 'musit/auth/LOGOUT';
-const LOGOUT_SUCCESS = 'musit/auth/LOGOUT_SUCCESS';
-const LOGOUT_FAIL = 'musit/auth/LOGOUT_FAIL';
+const SET_USER = 'musit/auth/SET_USER';
+const CLEAR_USER = 'musit/auth/CLEAR_USER';
 
 const initialState = {
-  loaded: false
+  user: null
 };
 
 const authReducer = (state = initialState, action = {}) => {
   switch (action.type) {
-    case LOAD:
+    case SET_USER:
       return {
         ...state,
-        loading: true
+        user: action.user
       };
-    case LOAD_SUCCESS:
+    case CLEAR_USER:
       return {
         ...state,
-        loading: false,
-        loaded: true,
-        user: action.result
-      };
-    case LOAD_FAIL:
-      return {
-        ...state,
-        loading: false,
-        loaded: false,
-        error: action.error
-      };
-    case LOGIN:
-      return {
-        ...state,
-        loggingIn: true
-      };
-    case LOGIN_SUCCESS:
-      return {
-        ...state,
-        loggingIn: false,
-        user: action.result
-      };
-    case LOGIN_FAIL:
-      return {
-        ...state,
-        loggingIn: false,
-        user: null,
-        loginError: action.error
-      };
-    case LOGOUT:
-      return {
-        ...state,
-        loggingOut: true
-      };
-    case LOGOUT_SUCCESS:
-      return {
-        ...state,
-        loggingOut: false,
         user: null
-      };
-    case LOGOUT_FAIL:
-      return {
-        ...state,
-        loggingOut: false,
-        logoutError: action.error
       };
     default:
       return state;
@@ -75,31 +24,15 @@ const authReducer = (state = initialState, action = {}) => {
 
 export default authReducer
 
-export const isLoaded = (globalState) => {
-  return globalState.auth && globalState.auth.loaded;
+export const connectUser = (user) => {
+  return {
+    type: SET_USER,
+    user: user
+  }
 }
 
-export const load = () => {
+export const clearUser = () => {
   return {
-    types: [LOAD, LOAD_SUCCESS, LOAD_FAIL],
-    promise: (client) => client.get('/loadAuth')
-  };
-}
-
-export const login = (name) => {
-  return {
-    types: [LOGIN, LOGIN_SUCCESS, LOGIN_FAIL],
-    promise: (client) => client.post('/login', {
-      data: {
-        name: name
-      }
-    })
-  };
-}
-
-export const logout = () => {
-  return {
-    types: [LOGOUT, LOGOUT_SUCCESS, LOGOUT_FAIL],
-    promise: (client) => client.get('/logout')
+    type: CLEAR_USER
   };
 }
