+++ gui/app/server.js
@@ -38,6 +38,11 @@ import { syncHistoryWithStore } from 'react-router-redux'
 import {Provider} from 'react-redux'
 import getRoutes from './routes'
 
+import Passport from 'passport'
+import {Strategy as DataportenStrategy} from 'passport-dataporten'
+import {Strategy as JsonStrategy} from 'passport-json-custom'
+import { connectUser } from './reducers/auth'
+
 
 const targetUrl = 'http://' + config.apiHost + ':' + config.apiPort
 const pretty = new PrettyError()
@@ -53,9 +58,77 @@ app.use(favicon(path.join(__dirname, '..', 'static', 'favicons', 'unimusfavicon.
 
 app.use(Express.static(path.join(__dirname, '..', 'static')))
 
+/* *** START Security *** */
+const dataportenUri = '/auth/dataporten'
+const dataportenCallbackUri = '/auth/dataporten/callback'
+const dataportenCallbackUrl = `http://${config.host}:8080/musit`
+var passportStrategy = null
+var passportLoginType = null
+console.log(dataportenCallbackUrl)
+
+if (config.FAKE_STRATEGY === config.dataportenClientSecret) {
+  // TODO:FAKEIT
+  console.log(' Installing strategy: LOCAL')
+  passportLoginType = 'json-custom'
+
+  const findUser = (username) => {
+    const securityDatabase = require('./fake_security.json')
+    return securityDatabase.users.find( (user) => user.userId == username)
+  }
+
+  const localCallback = (credentials, done) => {
+    var user = findUser(credentials.username)
+    if (!user) {
+      done(null, false, { message: 'Incorrect username.' })
+    } else {
+      done(null, user)
+    }
+  }
+
+  passportStrategy = new JsonStrategy( localCallback )
+} else {
+  // TODO: Consider placing this initialization strategy into the config object
+  passportLoginType = 'dataporten'
+  const dpConfig = {
+    clientID: config.dataportenClientID,
+    clientSecret: config.dataportenClientSecret,
+    callbackURL: dataportenCallbackUrl
+  }
+
+  const dpCallback = (accessToken, refreshToken, profile, done) => {
+    //load user and return done with the user in it.
+
+    // TODO: Add user info to redux state
+    return done(null, {
+      userId: profile.data.id,
+      name: profile.data.displayName,
+      emails: profile.data.emails,
+      photos: profile.data.photos,
+      accessToken: accessToken
+    })
+  }
+
+  passportStrategy = new DataportenStrategy(dpConfig, dpCallback)
+}
+
+Passport.serializeUser((user, done) => {
+  done(null, user);
+})
+
+Passport.deserializeUser((user, done) => {
+  done(null, user);
+})
+
+Passport.use(
+  passportStrategy
+)
+
+app.use(Passport.initialize())
+
 // Proxy to API server
 app.use('/api', (req, res) => {
   proxy.web(req, res, {target: targetUrl})
+
 })
 
 app.use('/ws', (req, res) => {
@@ -80,56 +153,70 @@ proxy.on('error', (error, req, res) => {
   res.end(JSON.stringify(json))
 })
 
-app.use((req, res) => {
-  if (__DEVELOPMENT__) {
-    // Do not cache webpack stats: the script file would change since
-    // hot module replacement is enabled in the development env
-    webpackIsomorphicTools.refresh()
-  }
-  const client = new ApiClient(req)
-  const virtualBrowserHistory = createHistory(req.originalUrl)
-
-  const store = createStore(client)
-
-  const history = syncHistoryWithStore(virtualBrowserHistory, store)
-
-  function hydrateOnClient() {
-    res.send('<!doctype html>\n' +
-      ReactDOM.renderToString(<Html assets={webpackIsomorphicTools.assets()} store={store}/>))
-  }
-
-  if (__DISABLE_SSR__) {
-    hydrateOnClient()
-    return
-  }
-
-  match({ history, routes: getRoutes(store), location: req.originalUrl }, (error, redirectLocation, renderProps) => {
-    if (redirectLocation) {
-      res.redirect(redirectLocation.pathname + redirectLocation.search)
-    } else if (error) {
-      console.error('ROUTER ERROR:', pretty.render(error))
-      res.status(500)
-      hydrateOnClient()
-    } else if (renderProps) {
-      loadOnServer({...renderProps, store, helpers: {client}}).then(() => {
-        const component = (
-          <Provider store={store} key="provider">
-            <ReduxAsyncConnect {...renderProps} />
-          </Provider>
-        )
-
-        res.status(200)
-
-        global.navigator = {userAgent: req.headers['user-agent']}
-
-        res.send('<!doctype html>\n' +
-          ReactDOM.renderToString(<Html assets={webpackIsomorphicTools.assets()} component={component} store={store}/>))
-      })
-    } else {
-      res.status(404).send('Not found')
-    }
-  })
+app.get('/', (req, res) => {
+  res.status(200).send('<!doctype html>\n<html>\n<body>\n<a href="/musit">Login</a>\n</body>\n</html>\n')
 })
+
+app.use('/musit', Passport.authenticate(passportLoginType, {failWithError: true}),
+  (req, res) => {
+      if (__DEVELOPMENT__) {
+        // Do not cache webpack stats: the script file would change since
+        // hot module replacement is enabled in the development env
+        webpackIsomorphicTools.refresh()
+      }
+      const client = new ApiClient(req)
+      const virtualBrowserHistory = createHistory(req.originalUrl)
+
+      const store = createStore(client, {auth: {user: req.user}})
+
+      const history = syncHistoryWithStore(virtualBrowserHistory, store)
+
+      function hydrateOnClient() {
+        res.send('<!doctype html>\n' +
+          ReactDOM.renderToString(<Html assets={webpackIsomorphicTools.assets()} store={store}/>))
+      }
+
+      if (__DISABLE_SSR__) {
+        hydrateOnClient()
+        return
+      }
+
+      store.dispatch(connectUser(req.user))
+
+      match({ history, routes: getRoutes(store), location: req.originalUrl }, (error, redirectLocation, renderProps) => {
+        if (redirectLocation) {
+          res.redirect(redirectLocation.pathname + redirectLocation.search)
+        } else if (error) {
+          console.error('ROUTER ERROR:', pretty.render(error))
+          res.status(500)
+          hydrateOnClient()
+        } else if (renderProps) {
+          loadOnServer({...renderProps, store, helpers: {client}}).then(() => {
+            const component = (
+              <Provider store={store} key="provider">
+                <ReduxAsyncConnect {...renderProps} />
+              </Provider>
+            )
+
+            res.status(200)
+
+            global.navigator = {userAgent: req.headers['user-agent']}
+
+            res.send('<!doctype html>\n' +
+              ReactDOM.renderToString(<Html assets={webpackIsomorphicTools.assets()} component={component} store={store}/>))
+          })
+        } else {
+          res.status(404).send('Not found')
+        }
+      })
+  },
+  (err, req, res, next) => {
+    res.status(400).json({
+      authenticated: req.isAuthenticated(),
+      err: err.message
+    })
+  }
+)
 
 if (config.port) {
   server.listen(config.port, (err) => {
